+++ src/identification/adjustment-mag.jl
@@ -231,12 +231,12 @@ function is_valid_adjustment(
     cg::AbstractAG,
     x::Union{Symbol,AbstractVector{Symbol}},
     y::Union{Symbol,AbstractVector{Symbol}},
-    z::AbstractVector{Symbol} = Symbol[],
+    z::Union{Symbol,AbstractVector{Symbol}} = Symbol[],
 )
     B = cg.backend
     xs = _node_indices(cg, x)
     ys = _node_indices(cg, y)
-    z_idxs = [node_index(cg, v) for v in z]
+    z_idxs = _node_indices(cg, z)
 
     forbidden = _forbidden_set(B, xs, ys)
     any(v -> forbidden[v], z_idxs) && return false